fix: 회원 클래스 병합
-겹치는 클래스 병합
